feat: add ui/slider, rework process to steps component, rework slider component to process
