+++ landing/fragments/landing-process/src/slider/slider.images.ts
@@ -1,0 +1,2 @@
+export const imagesDesktop = ['./image/mockup.png', './image/mockup-negate.png']
+export const imagesMobile = ['./image/mockup.png', './image/mockup-negate.png']
