+++ landing/fragments/landing-steps/package.json
@@ -1,12 +1,11 @@
 {
-  "name": "@landing/fragment-slider",
+  "name": "@landing/fragment-steps",
   "version": "0.0.1",
   "main": "src/index.ts",
   "dependencies": {
     "@ui/background": "workspace: 0.0.1",
-    "@ui/button": "workspace: 0.0.1",
     "@ui/condition": "workspace: 0.0.1",
-    "@ui/icons": "workspace: 0.0.1",
+    "@ui/divider": "workspace: 0.0.1",
     "@ui/image": "workspace: 0.0.1",
     "@ui/layout": "workspace: 0.0.1",
     "@ui/text": "workspace: 0.0.1",
