+++ ui/accordion/package.json
@@ -15,7 +15,6 @@
     "@emotion/css": "11.11.0",
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
-    "@types/dompurify": "2.4.0",
     "@types/styled-system": "5.1.16",
     "react": "18.2.0",
     "react-dom": "18.2.0",
