+++ ui/image/src/image.interfaces.ts
@@ -1,4 +1,5 @@
 export interface ImageProps {
+  hover?: boolean
   width: number
   height: number
 }
