+++ ui/slider/package.json
@@ -1,13 +1,18 @@
 {
-  "name": "@landing/fragment-process",
+  "name": "@ui/slider",
   "version": "0.0.1",
+  "license": "BSD-3-Clause",
   "main": "src/index.ts",
   "dependencies": {
+    "@ui/background": "workspace: 0.0.1",
+    "@ui/button": "workspace: 0.0.1",
     "@ui/condition": "workspace: 0.0.1",
+    "@ui/icons": "workspace: 0.0.1",
+    "@ui/image": "workspace: 0.0.1",
     "@ui/layout": "workspace: 0.0.1",
-    "@ui/slider": "workspace: 0.0.1",
     "@ui/text": "workspace: 0.0.1",
-    "@ui/utils": "workspace: 0.0.1"
+    "@ui/utils": "workspace: 0.0.1",
+    "framer-motion": "^10.16.1"
   },
   "devDependencies": {
     "@types/react": "18.2.9",
