+++ ui/slider/src/index.ts
