+++ ui/slider/src/slider.interfaces.ts
@@ -1,0 +1,3 @@
+export interface SliderProps {
+  images: string[]
+}
